Improve editor selection listener
The selection listener that tries to highlight fields in the java editor did not
take into account the fact that Eclipse deactivates parts if they are idle for
longer periods of time. When the listener looks for the java editor in the editor references,
it should do a null check on the editor itself, before examining it further. It should
also try to restore it when getting the reference, which is achieved by calling
IEditorReference.getEditor() with it's boolean parameter set to true.

Another slight improvement is that the listener should only bother to try to select
fields if the java editor is visible in the workbench page.