Try to get the document listener from the context
Since DI isn't supported in the normal case in 3.x, this is probably the only way to decouple the validator and the document listener, in a manner suitable for unit testing. It is quite hacky. In a production enviroment, the listener will NOT be present in the context, and the validator must acquire it from the activator, instead. Under testing, a (mocked) listener can be put in the context prior to instantiating the validator.