Use Maven to get Scene Builder's dependencies
Get Scene Builder's dependencies from Maven Central Instead of manually including them. This involves using the Maven bundle plugin to generate a bundle that wraps the dependencies. Note that the wrapper bundle needs to be built separately. The Gluon dependencies are not included at this point, since they depend on JavaFX at both build and runtime.