Synchronize external changes
Add a resource change listener that tracks changes made to the editor's document from outside of Eclipse itself, so that the editor is in sync with the underlying file at all times. Note that this works best when the "Refresh using native hooks or pulling" option is enabled; otherwise, a refresh of the workspace is required to detect the changes.