Add custom builder
Add a custom project nature along with a builder for it that performs the coherence validation. It is better to perform validation from a builder rather than a compilation participant as it receives notification about resources as well. Hence, it is the easiest way to detect all three discrepancy types from a single point and mechanism. A drawback is that marker resolutions have to be used instead of quick fixes that can utilize the richer completion proposals (and ASTRewrites). In particular, they create previews automatically.